refactor: migrate window management and later-callback utilities from utils.ts to lib.ts
